+++ src/features/auth/components/FormContainer/index.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { View } from 'react-native';
+
+import styles from './styles';
+import { FormContainerProps } from './types';
+
+export const FormContainer: React.FunctionComponent<FormContainerProps> = ({ children }) => (
+  <View style={styles.mainContainer}>{children}</View>
+);
