+++ src/features/auth/components/FormContainer/styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+
+import { WHITE } from 'constants/styles';
+
+export default StyleSheet.create({
+  mainContainer: {
+    alignItems: 'center',
+    borderRadius: 8,
+    backgroundColor: WHITE,
+    marginHorizontal: 24,
+    marginVertical: 8,
+    paddingHorizontal: 28,
+    paddingVertical: 20,
+  },
+});
