+++ src/features/auth/components/FormContainer/types.ts
@@ -1,0 +1,3 @@
+export type FormContainerProps = {
+  children: React.ReactNode;
+};
