+++ src/features/auth/welcome/screen/styles.ts
@@ -1,17 +1,8 @@
 import { StyleSheet } from 'react-native';
 
-import { RED, WHITE } from 'constants/styles';
+import { RED } from 'constants/styles';
 
 export default StyleSheet.create({
-  mainContainer: {
-    alignItems: 'center',
-    borderRadius: 8,
-    backgroundColor: WHITE,
-    marginHorizontal: 24,
-    marginVertical: 8,
-    paddingHorizontal: 28,
-    paddingVertical: 20,
-  },
   logo: {
     resizeMode: 'contain',
     width: 173,
